+++ services/api/business/builder/accounts.go
@@ -1,15 +1,16 @@
 package builder
 
 import (
-	qkmclient "github.com/consensys/quorum-key-manager/pkg/client"
 	usecases "github.com/consensys/orchestrate/services/api/business/use-cases"
 	"github.com/consensys/orchestrate/services/api/business/use-cases/accounts"
 	"github.com/consensys/orchestrate/services/api/store"
+	qkmclient "github.com/consensys/quorum-key-manager/pkg/client"
 )
 
 type accountUseCases struct {
 	createAccountUC  usecases.CreateAccountUseCase
 	getAccountUC     usecases.GetAccountUseCase
+	deleteAccountUC  usecases.DeleteAccountUseCase
 	searchAccountsUC usecases.SearchAccountsUseCase
 	updateAccountUC  usecases.UpdateAccountUseCase
 }
@@ -26,6 +27,7 @@ func newAccountUseCases(
 
 	return &accountUseCases{
 		createAccountUC:  accounts.NewCreateAccountUseCase(db, searchAccountsUC, fundAccountUC, keyManagerClient),
+		deleteAccountUC:  accounts.NewDeleteAccountUseCase(db, keyManagerClient),
 		getAccountUC:     accounts.NewGetAccountUseCase(db),
 		searchAccountsUC: searchAccountsUC,
 		updateAccountUC:  accounts.NewUpdateAccountUseCase(db),
@@ -46,4 +48,8 @@ func (u *accountUseCases) CreateAccount() usecases.CreateAccountUseCase {
 
 func (u *accountUseCases) UpdateAccount() usecases.UpdateAccountUseCase {
 	return u.updateAccountUC
+}
+
+func (u *accountUseCases) DeleteAccount() usecases.DeleteAccountUseCase {
+	return u.deleteAccountUC
 }